tournament-schedule: Refactor using the TournamentMatch component
